Renamed object classes to avoid conflict. Created ObjectTypeEnum
