+++ MonkeyInterpreter.Core/ObjectTypeEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace MonkeyInterpreter.Core;
+
+public enum ObjectTypeEnum
+{
+	Null,
+	Integer,
+	Boolean,
+}
